Application - API/requirements.txt
gunicorn==20.1.0
Flask==1.1.2
Flask-RESTful==0.3.9
google-api-core==2.0.1
google-api-python-client==2.20.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==2.0.0
google-cloud-storage==1.42.1
grpcio==1.34.0
tensorflow==2.5.0
transformers==4.9.1
